Updated Code for IR Module
The previous code turned off one motor and rotated the other to make a turn, which caused the vehicle to overshoot the line. I modified the logic to reverse the other wheel instead of stopping completely during turns to improve stability in line following. This change helps maintain better control while navigating.